requirements.txt
yandex-cloud-ml-sdk
pydantic~=2.11.3
python-dotenv~=1.1.0
pandas
telebot
yandex-speechkit

